Breaks out some submodules.
