+++ mercutio-server/src/bin/mercutio-server.rs
@@ -1,11 +1,11 @@
 //! mercutio-server standalone binary for web deployment.
 
+#![feature(extern_in_paths)]
 #![feature(proc_macro)]
 #![feature(proc_macro_non_items)]
 
 extern crate bus;
 extern crate crossbeam_channel;
-#[macro_use] extern crate failure;
 extern crate include_dir_macro;
 extern crate maplit;
 extern crate mercutio_common;
@@ -29,7 +29,6 @@ extern crate md5;
 #[macro_use]
 extern crate serde_json;
 
-use failure::Error;
 use include_dir_macro::include_dir;
 use mercutio_common::{
     doc_as_html,
@@ -38,7 +37,12 @@ use mercutio_common::{
         doc_to_markdown,
     },
 };
-use mercutio_server::sync::*;
+use extern::{
+    mercutio_server::{
+        graphql::client::*,
+        sync::*,
+    },
+};
 use mime_guess::guess_mime_type;
 use oatie::doc::*;
 use oatie::validate::validate_doc;
@@ -125,108 +129,6 @@ Type github.com/tcr/edit-text into your search bar for more information.
     let doc = Doc(markdown_to_doc(&INPUT).unwrap());
     validate_doc(&doc).expect("Initial Markdown document was malformed");
     doc
-}
-
-pub fn get_single_page_graphql(input_id: &str) -> Option<Doc> {
-    let client = reqwest::Client::new();
-    let text = client.post("http://127.0.0.1:8003/graphql/")
-        .json(&json!({
-            "query": r#"
-
-query ($id: String!) {
-    page(id: $id) {
-        doc
-    }
-}
-
-"#,
-            "variables": {
-                "id": input_id,
-            },
-        }))
-        .send()
-        .ok()?
-        .text()
-        .ok()?;
-    
-    let ret: ::serde_json::Value = serde_json::from_str(&text).ok()?;
-    let node = ret.pointer("/data/page/doc")?;
-    let ron = node.as_str()?.to_string();
-    let body = ::ron::de::from_str(&ron).ok()?;
-    Some(Doc(body))
-}
-
-pub fn graphql_request(
-    query: &str,
-    variables: &serde_json::Value,
-) -> Result<serde_json::Value, Error> {
-    let client = reqwest::Client::new();
-    let text = client.post("http://127.0.0.1:8003/graphql/")
-        .json(&json!({
-            "query": query,
-            "variables": variables,
-        }))
-        .send()?
-        .text()?;
-    
-    // TODO handle /errors[...]
-    Ok(serde_json::from_str(&text)?)
-}
-
-pub fn get_or_create_page_graphql(input_id: &str, doc: &Doc) -> Result<Doc, Error> {
-    let ret = graphql_request(
-        r#"
-
-mutation ($id: String!, $default: String!) {
-    getOrCreatePage(id: $id, default: $default) {
-        doc
-    }
-}
-
-"#,
-        &json!({
-            "id": input_id,
-            "default": ::ron::ser::to_string(&doc.0).unwrap(),
-        }),
-    )?;
-
-    // Extract the doc field.
-    let doc_string = ret.pointer("/data/getOrCreatePage/doc")
-        .ok_or(format_err!("unexpected json structure"))?
-        .as_str().unwrap()
-        .to_string();
-
-    Ok(Doc(::ron::de::from_str(&doc_string)?))
-}
-
-pub fn create_page_graphql(input_id: &str, doc: &Doc) -> Option<Doc> {
-    let client = reqwest::Client::new();
-    let text = client.post("http://127.0.0.1:8003/graphql/")
-        .json(&json!({
-            "query": r#"
-
-mutation ($id: String!, $doc: String!) {
-    createPage(id: $id, doc: $doc) {
-        doc
-    }
-}
-
-"#,
-            "variables": {
-                "id": input_id,
-                "doc": ::ron::ser::to_string(&doc.0).unwrap(),
-            },
-        }))
-        .send()
-        .ok()?
-        .text()
-        .ok()?;
-    
-    let ret: ::serde_json::Value = serde_json::from_str(&text).ok()?;
-    let node = ret.pointer("/data/createPage/doc")?;
-    let ron = node.as_str()?.to_string();
-    let body = ::ron::de::from_str(&ron).ok()?;
-    Some(Doc(body))
 }
 
 fn run_http_server(port: u16, client_proxy: bool) {
